IOS fixes, CurrentPlayInfo separate component
