+++ src/webapp/features/rabble/Buttons.module.css
@@ -13,6 +13,11 @@
   margin-left: 2px;
   margin-right: 2px;
   padding: 2px;
+  min-height: 40px;
+  -webkit-appearance: none;
+  background-color: rgb(239, 239, 239);
+  border-radius: 3px;
+  border-width: 1px;
 }
 
 .buttons > button > svg {
@@ -22,6 +27,7 @@
 
 .play {
   color: green;
+
   box-shadow: 1px 1px 3px -1px dimgray;
   font-weight: bold;
 }
