+++ src/webapp/features/rabble/components/SwapTiles.module.css
@@ -27,6 +27,10 @@
   width: 45px;
   height: 45px;
   font-size: x-small;
+  -webkit-appearance: none;
+  border-radius: 3px;
+  border-width: 1px;
+  background-color: rgb(239, 239, 239);
 }
 
 .topRow > button > svg {
